Moved ConnectionPanel MVC to connection.py
